requirements.txt
mesop>=0.0.40
psycopg2-binary
python-dotenv
